+++ commands/src/apply/mod.rs
@@ -1,5 +1,6 @@
+pub mod command;
 pub mod config;
-pub mod init;
 pub mod lockfile;
 pub mod metrics;
+pub mod report;
 pub mod types;